hubble/parser: Always preserve datapath numeric identity
This introduces a check that we do not overwrite the numeric security
identity provided by the datapath trace point. Only if the datapath did
not provide an identity (i.e. in `FROM_LXC` trace points) do we want to
fall back on the identity from the user-space IPCache.

The datapath identity and the identity returned by `LookupSecIDByIP` may
differ if the identity has changed between the time the trace event has
been generated and the time Hubble observes it. By preserving what the
datapath originally observed, we ensure that we have the same
information as the datapath had when it made the policy decision for
this flow.

Signed-off-by: Sebastian Wicki <sebastian@isovalent.com>